Remove hard-coded greeting
In an effort to make this script as general as possible, I've
removed the Aweber-friendly greeting that adds the subscriber's
name.

Instead, I'll simply add that greeting to the source messages.

Small additional change: removed development code of wrapping
plaing text emails. Linus says that it's best to do this
where you have control over it, and that would be in the
source messages.