Bugfix: tag URLs in HTML
URLs where not being tagged correctly, as I was processing the
HTML text, but should have tagged the URLs when it was still
in Markdown format.

There's also a feature addition in this commit. By tagging
any URL (or anything for that matter) with the anchor
URL parameters for Google Analytics.