Bugfix: correctly extract campaign from filename
The campaign tag for Google Analytics is extracted from the name
of the given file, minus the extention. This worked fine when
the file provided was in the same directory of the script.

However, I found out that calling the script on a file, giving
an absolute path to that file, I was incorreclty assuming
I was extracting only the basename of the file. This has now
been corrected.